+++ docs/misc/changelog.rst
@@ -26,6 +26,8 @@ New Features:
 - Added ``monitor_kwargs`` parameter to ``make_vec_env`` and ``make_atari_env``
 - Wrap the environments automatically with a ``Monitor`` wrapper when possible.
 - ``EvalCallback`` now logs the success rate when available (``is_success`` must be present in the info dict)
+- Added ``render`` argument to ``SubprocVecEnv`` and ``DummyVecEnv`` enabling
+  rendering of terminal steps.
 
 Bug Fixes:
 ^^^^^^^^^^